chore: hooks / aligning tests suites
