+++ packages/integration-tests/package.json
@@ -10,6 +10,7 @@
   "scripts": {
     "postinstall": "cd src/playbook && npm install",
     "integration-test:api": "jest --verbose --testPathPattern=tokens.test.ts && jest --verbose --testPathPattern=deposit.test.ts && jest --verbose --testPathPattern=common && jest --verbose --testPathPattern=transactions",
+    "hooks:global": "jest --verbose --testPathPattern=global.test.ts",
     "integration-test:ui": "npx playwright test",
     "block-explorer:start": "docker-compose up",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-pattern 'repos' --ignore-pattern 'git add .'"
