+++ packages/integration-tests/tests/api/stats.test.ts
@@ -1,8 +1,8 @@
 import * as request from "supertest";
 import { setTimeout } from "timers/promises";
 
-import { environment } from "../../../../src/config";
-import { localConfig } from "../../../../src/config";
+import { environment } from "../../src/config";
+import { localConfig } from "../../src/config";
 
 describe("/stats", () => {
   jest.setTimeout(localConfig.standardTimeout);
