+++ packages/integration-tests/tests/hooks/global.test.ts
@@ -1,8 +1,29 @@
-import { localConfig } from "../../../src/config";
-import { Buffer } from "../../../src/entities";
-import { Logger } from "../../../src/entities";
-import { Helper } from "../../../src/helper";
-import { Playbook } from "../../../src/playbook/playbook";
+import { localConfig } from "../../src/config";
+import { Buffer, Logger } from "../../src/entities";
+import { Helper } from "../../src/helper";
+import { Playbook } from "../../src/playbook/playbook";
+
+describe("Tokens", () => {
+  jest.setTimeout(localConfig.standardTimeout);
+
+  let deployedToken: string;
+
+  describe("Deploy/check the custom ERC20 tokens", () => {
+    const playbook = new Playbook();
+
+    //@id603
+    it("Deploy custom ERC20 token to L2", async () => {
+      deployedToken = await playbook.deployERC20toL2();
+      expect(deployedToken).toContain(Logger.txHashStartsWith);
+    });
+
+    //@id657
+    it("Deploy custom ERC20 token to L1", async () => {
+      deployedToken = await playbook.deployERC20toL1();
+      expect(deployedToken).toContain(Logger.txHashStartsWith);
+    });
+  });
+});
 
 describe("Deposit", () => {
   const playbook = new Playbook();
