+++ packages/integration-tests/tests/hooks/local/multiCall.test.ts
@@ -1,6 +1,6 @@
-import { localConfig } from "../../../../../src/config";
-import { Logger } from "../../../../../src/entities";
-import { Playbook } from "../../../../../src/playbook/playbook";
+import { localConfig } from "../../../src/config";
+import { Logger } from "../../../src/entities";
+import { Playbook } from "../../../src/playbook/playbook";
 
 describe("Multicall transactions", () => {
   jest.setTimeout(localConfig.extendedTimeout);
